LPS-179567 Extract CommercePricingClassProductsScreenNavigationCategory methods to a CommercePricingClassProductsScreenNavigationEntry component and extends CommercePricingClassProductsScreenNavigationCategory in it
